Add gaussian blur of each color
This needs another tempBuf, so now there are tempBuf1 and tempBuf2.
Doing the blur in place affects nearby pixels that shoudln't be
affected, and reading from frontBuf is slow. For this, read from
tempBuf2 to make the blur in tempBuf1, then copy tempBuf1 to backBuf,
and then swap tempBuf1 and tempBuf2.