Switch the order of compute and graphics
For some reason, graphics would ALWAYS seem to run after compute, even
with the semaphore, fence, and memory barrier in the middle. In fact,
setting the graphics load operation to CLEAR instead of DONT_CARE would
actually clear the compute output that was supposed to happen after the
whole graphics renderpass, showing that somehow it may have been
deferred until the last moment.

To work around this I switched the order of compute and graphics. Now
compute goes first, and writes a pixel with alpha 0.5. The fragment
shader will then read the back image, which is now included as an input
attachment, and if the alpha is 0.5, write the particle color with alpha
1.0 and return. Otherwise it reads the front image and (not yet) applies
the blur and fade.